refactor: Separate utility functions
