+++ electron/src/util/OriginValidator.ts
@@ -1,0 +1,9 @@
+import * as url from 'url';
+
+class OriginValidator {
+  static isMatchingHost(urlString: string, baseUrl: string): boolean {
+    return url.parse(urlString).host === url.parse(baseUrl).host;
+  }
+}
+
+export {OriginValidator};
